J-274: +styles-import+shadow
