+++ src/screens/PodcastPage/PodcastPage.tsx
@@ -8,6 +8,7 @@ import {
   ImageBackground
 } from 'react-native'
 import EpisodeElement from './components/episodeBlock'
+import styles from './styles'
 import { ScrollView } from 'react-native-gesture-handler'
 
 const win = Dimensions.get('window')
@@ -33,11 +34,13 @@ const PodcastPage: React.FC = () => {
         >
           <View style={styles.backButton}>
           </View>
+          <View style={styles.podcastLogoContainer}>
           <Image
             source={imagepodcast}
             resizeMode="cover"
             style={styles.podcastLogo}
           ></Image>
+          </View>
         </ImageBackground>
       </View>
       <View style={styles.podcastNameContainer}>
@@ -64,67 +67,5 @@ const PodcastPage: React.FC = () => {
     </ScrollView>
   )
 }
-
-const styles = StyleSheet.create({
-  podcastBackground: {
-    flex: 1,
-    height: 150,
-    width: win.width,
-    marginBottom: 65
-  },
-  podcastLogo: {
-    flex: 1,
-    height: 120,
-    width: 120,
-    position: 'absolute',
-    alignSelf: 'center',
-    bottom: -60,
-    borderRadius: 5
-  },
-  podcastNameContainer: {
-    marginTop: 5,
-    alignItems: 'center'
-  },
-  podcastName: {
-    textAlign: 'center',
-    fontSize: 24,
-    width: 250,
-    fontWeight: 'bold'
-  },
-  podcastAuthorText: {
-    color: '#000',
-    fontSize: 15,
-  },
-  description: {
-    alignItems: 'center'
-  },
-  descriptionText: {
-    textAlign: 'center',
-    width: 300,
-    marginTop: 7,
-    lineHeight: 20,
-    color: '#555'
-  },
-  episodeCounter: {
-    alignItems: 'center',
-    marginTop: 20,
-  },
-  episodesContainer: {
-    marginTop: 20,
-    paddingHorizontal: 35,
-  },
-  episodesContainerTitle: {
-    fontSize: 20,
-    color: '#444',
-    fontWeight: 'bold',
-    marginBottom: 20
-  },
-  backButton: {
-      backgroundColor: '#aaa',
-      width: 45,
-      height: 45,
-      margin: 10,
-  }
-})
 
 export default PodcastPage
